requirements.txt
transformers==4.56.1
datasets==4.1.0
bitsandbytes==0.47.0
accelerate==1.10.1
torch==2.8.0
langchain==0.3.27
langchain-openai==0.3.33
langchain-huggingface==0.3.1
python-dotenv==1.1.1
rich==14.1.0